Added new test - renamed performance tests to concurrency - removed unnecessary test cases - added new concurrency test case
