+++ src/__tests__/methods.test.ts
@@ -1,4 +1,4 @@
-/* 
+/*
 eth-address-observer is free software: you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
 the Free Software Foundation, either version 3 of the License, or
